Fix incorrect int parsing during is_numeric checks
Addresses https://github.com/reubano/meza/issues/36

Where a zero-valued currency string e.g. '$0' or '0$' is passed to
`is_numeric`, the *stripped* value should be coerced to an int when
testing if the value can be parsed to integer 0. Previously, the
unstripped value was being coerced which leads to a `ValueError` being
raised.

Signed-off-by: Ben Steadman <steadmanben1@gmail.com>